Repackage lint into a standalone jar
The lint shell script in tools/bin/lint uses tools/lib/lint-classpath.jar
as its classpath.  That jar contains a list of other jars that java adds
to the classpath.  This makes it impossible for the build system to track
dependencies.

Use a genrule to read the jars from tools/lib/lint-classpath.jar and then
merge them all together into a single jar.  The result is then embedded
into a self-contained java_binary.

Bug: 181681346
Test: m lint && lint --help
Change-Id: Iea1272200b3578b0f7bd3ad8c985e7a38fc4cf09